refactor the refresh logic into the AuthManager
Makes more sense here tbh
